requirements.txt
python-dotenv
requests
flask
python-dotenv
google-generativeai